refactor: 컨벤션을 위한 코드 수정 작업
