+++ frontend/src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as Loading } from './Loading/Loading';
+export { default as Form } from './Form/Form';
