--- frontend/src/pages/SignUp/SignUp.styles.tsx
@@ -1,1 +1,0 @@
-export {};
